WIP decoupled all common code between kucoin and binance to ExchangeBaseV2
