+++ hummingbot/connector/constants.py
@@ -1,0 +1,5 @@
+NaN = float("nan")
+s_decimal_NaN = Decimal("nan")
+s_decimal_0 = Decimal(0)
+MINUTE = 60
+TWELVE_HOURS = MINUTE * 60 * 12
